added auto-gen policy rule for cli
