+++ pkg/common/common.go
@@ -1,0 +1,7 @@
+package common
+
+// Policy Reporting Modes
+const (
+	Enforce = "enforce" // blocks the request on failure
+	Audit   = "audit"   // dont block the request on failure, but report failiures as policy violations
+)
